% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotators.R
\docType{methods}
\name{annotateAnchors}
\alias{annotateAnchors}
\alias{annotateAnchors,GenomicInteractions,numeric,character,vector-method}
\title{Annotate anchors - DEPRECATED}
\usage{
annotateAnchors(GIObject, oneOrTwo, name, dat)


  \S4method{annotateAnchors}{GenomicInteractions,numeric,character,vector}(GIObject,
  oneOrTwo, name, dat)
}
\arguments{
\item{GIObject}{A GenomicInteractions object}

\item{oneOrTwo}{An integer indicating which anchor to annotate}

\item{name}{Character. Will be used as a column name for the elementMetadata 
of the annotated anchor.}

\item{dat}{Vector of the same length as the GenomicInteractions object,
containing data with which to annotate the object.}
}
\value{
invisible(1)
}
\description{
This function is deprecated and will be removed in the next release of Bioconductor.
Use `annotateRegions` instead.
}

