% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundingBox.R
\docType{methods}
\name{boundingBox}
\alias{boundingBox}
\title{Get the bounding box}
\usage{
\S4method{boundingBox}{GenomicInteractions}(x, f, reflect = TRUE)
}
\arguments{
\item{x}{A \linkS4class{GenomicInteractions} object.}

\item{f}{A factor or vector of length equal to that of \code{x}, 
indicating the group to which each pairwise interaction belongs.}

\item{reflect}{A logical scalar indicating whether intra-chromosomal interactions can be flipped around the diagonal to achieve a smaller bounding box.}
}
\value{
A \linkS4class{GenomicInteractions} object containing the coordinates of each bounding box.
}
\description{
Compute a minimum bounding box for groups of pairwise genomic interactions.
}
\details{
For any group of pairwise interactions, the minimum bounding box is the smallest rectangle in the interaction space that contains all interactions in the group.
Each side of the box has coordinates spanning the most extreme anchor regions on the corresponding chromosome.
This is often useful for summarizing clusters of interactions.

Grouping of interactions is specified using \code{f}, where interactions in \code{x} with the same level of \code{f} are considered to be in the same group.
If \code{f} is not specified, all interactions in \code{x} are assumed to be in a single group (named as ``1'').
An error will be raised if a group spans multiple chromosomes for either the first or second anchor regions.

The function returns a \linkS4class{GenomicInteractions} object containing the coordinates of the bounding boxes for all groups.
Each interaction represents a bounding box for a group, where the anchor regions represent the sides of the box.
Entries are named according to the levels of \code{f}, in order to specify which bounding box corresponds to which group.

For intra-chromosomal groups, \code{reflect=TRUE} will ensure that all interactions lie on one side of the diagonal of the intra-chromosomal interaction space.
(Specifically, the first anchor will always start before the second anchor.)
This yields a minimum bounding box with the smallest possible larger dimension,
which will only increase in size if interactions are placed on the other side of the diagonal.

% Bit of a pain to prove, but basically, if you flip a point to be above the diagonal, the Chebyshev distance to a point below the diagonal will always increase.
% This means that you must increase the size of one of your sides of your bounding box.
}
\examples{
anchor1 <- GRanges(c("chr1", "chr1", "chr1", "chr1"), 
    IRanges(c(10, 20, 30, 20), width=5))
anchor2 <- GRanges(c("chr1", "chr1", "chr1", "chr2"), 
    IRanges(c(100, 200, 300, 50), width=5))
gi <- GenomicInteractions(anchor1, anchor2)

# Making up a sensible grouping, e.g., chromosome pairs.
f <- c(1,1,1,2)
boundingBox(gi, f)

# Fails for multiple chromosomes
try(out <- boundingBox(gi))

}
\author{
Aaron Lun
}
