% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize first letter of string}
\usage{
capitalize(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
a string with the first letter capitalised
}
\description{
This function will capitalize the first letter of each string in
a character vector, and lowercase following letters.
}

