% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversion}
\alias{conversion}
\alias{coerce,GenomicInteractions,Pairs-method}
\alias{makeGInteractionsFromGRangesPairs}
\title{Converting to/from pairs}
\usage{
makeGInteractionsFromGRangesPairs(x)
}
\arguments{
\item{x}{A Pairs object with two \linkS4class{GenomicRanges} entries.}
}
\value{
An object of the converted type.
}
\description{
Convert a \linkS4class{GenomicInteractions} object to a \linkS4class{Pairs} object, or vice versa.
}
\details{
An automated coercion from a Pairs object to a GenomicInteractions object is not possible,
as the former may not contain two GenomicRanges objects (in which case no standard conversion exists).
Hence the need for the more wordy function to be explicit about the expected inputs.
}
\examples{
anchor1 <- GRanges(c("chr1", "chr1", "chr1", "chr1"), 
    IRanges(c(10, 20, 30, 20), width=5))
anchor2 <- GRanges(c("chr1", "chr1", "chr1", "chr2"), 
    IRanges(c(100, 200, 300, 50), width=5))
test <- GenomicInteractions(anchor1, anchor2)

p <- as(test, "Pairs")
p

makeGInteractionsFromGRangesPairs(p)

}
\author{
Aaron Lun
}
