% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.importHicLib}
\alias{.importHicLib}
\title{Function to read in processed Hi-C interaction data generated by hiclib}
\usage{
.importHicLib(dir, chr_names)
}
\arguments{
\item{dir}{directory containing the output files generated by hiclib, extracted using h5dictToTxt.py}

\item{chr_names}{ordered chromosome names (for renaming)}
}
\value{
data.table containing information on the individual interactions
}
\description{
This function reads in the interaction data processed by [hiclib](https://bitbucket.org/mirnylab/hiclib). The data is extracted from hiclib and
exported from its internal HDF5 format to text using h5dictToTxt.py. This function reads in the relevant files present in the
directory and generates a data.table containing all of the information. It assumes that there are files with the following
names in the directory: fragids1, chrms1, mids1, fraglens1, fragids2, chrms2, mids2, fraglens2, distances.
}
