% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.importHomer}
\alias{.importHomer}
\title{Function to read in processed Hi-C interaction data generated by HOMER}
\usage{
.importHomer(fn)
}
\arguments{
\item{fn}{location of data exported by HOMER}
}
\value{
a data frame containing the relevant information
}
\description{
This function reads in the interaction data outputted by HOMER (http://homer.salk.edu/homer/interactions/).
}
