% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.readBam}
\alias{.readBam}
\title{Function to read in interaction-data stored in a BAM file}
\usage{
.readBam(fn)
}
\arguments{
\item{fn}{name of BAM file containing interaction information}
}
\value{
list of GRanges - storing the anchor information for each interaction
}
\description{
Reads in interactions stored in a BAM file. Assumes that each interaction is represented by pair of PETs
with the same qname. The function reads in and determines which anchor is which by examining the
isFirstMateRead and isSecondMateRead fields in the BAM file.
}
