% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.validateInput}
\alias{.validateInput}
\title{Function to validate tabular input}
\usage{
.validateInput(x, h)
}
\arguments{
\item{x}{A data frame.}

\item{h}{A character vector containing expected headings}
}
\value{
list of two GRanges, storing the anchor information for each interaction
}
\description{
Check for columns in a data.frame.
}
