% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{duplicated,GenomicInteractions-method}
\alias{duplicated,GenomicInteractions-method}
\title{duplicated, GenomicInteractions-method}
\usage{
\S4method{duplicated}{GenomicInteractions}(x, fromLast = FALSE,
  dropMetadata = FALSE)
}
\arguments{
\item{x}{A GenomicInteractions object}

\item{fromLast}{Whether to identify duplicates starting from last item in the
Genomicinteractions object or not. Default: FALSE.}

\item{dropMetadata}{Logical, default FALSE. Whether to drop interaction mcols
when considering unique interactions.}
}
\value{
A vector containing indices of duplicated interactions
}
\description{
Finds duplicated interactions in a GenomicInteractions object.
}
\details{
Uniqueness is based on anchor positions and metadata, interaction counts, and
interaction metadata.
}

