% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{export.chiasig}
\alias{export.chiasig}
\alias{export.chiasig,GenomicInteractions-method}
\title{Export interactions in a BEDPE-like format for use with ChiaSig}
\usage{
export.chiasig(GIObject, fn = NULL, score = "counts")

\S4method{export.chiasig}{GenomicInteractions}(GIObject, fn = NULL,
  score = "counts")
}
\arguments{
\item{GIObject}{A GenomicInteractions object.}

\item{fn}{A filename to write the interactions data to}

\item{score}{Which metadata column to use as the score: counts or normalised}
}
\value{
invisible(1) if outputting to file or a data.frame containing all of the corresponding information
}
\description{
Exports a GenomicInteractions object to BEDPE like format, (anchor specifications and a column for reads connecting them)
and writes to a specified file. If filename is not specified,
then a data.frame containing the information is returned. The value of the score parameter defines which field is used
to populate the score field.
}

