% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GenomicInteractions-class}
\alias{GenomicInteractions-class}
\title{A S4 class to represent interactions between genomic regions.}
\description{
A S4 class to represent interactions between genomic regions.
}
\section{Slots}{

\describe{
\item{\code{experiment_name}}{Character. Experiment name.}

\item{\code{description}}{Character. Longer description of experiment.}

\item{\code{genome_name}}{Character. Genome version for experiment data, should correspond to a BSgenome data package.}

\item{\code{anchor_one,anchor_two}}{GRanges. Set of anchors of interactions.}

\item{\code{counts}}{Numeric. Counts of reads supporting each interaction.}

\item{\code{normalised_counts}}{Numeric. Normalised counts of reads supporting each interaction.}

\item{\code{pvalue}}{Numeric. P-values for individual interactions.}

\item{\code{fdr}}{Numeric. FDRs for individual interactions.

This class is used to store information on which genomic regions are interacting with each other, the number of counts supporting each interaction,
and associated p-values and false-discovery rates (FDR). Objects of this class contain information of the genomic coordinates of the interacting
regions and the strength of these interactions, and associated metadata such as the name of the dataset and a brief description of the dataset.
Interacting regions are stored as a pair of GenomicRanges: each set of anchor regions is stored as a separate GenomicRanges object, accessed by
\code{getAnchorOne} and \code{getAnchorTwo}.}
}}
\examples{
showClass("GenomicInteractions")

library(BSgenome.Mmusculus.UCSC.mm9)
anchor.one = GRanges(c("chr1", "chr1", "chr1", "chr1"), IRanges(c(10, 20, 30, 20), width=5), seqlengths=seqlengths(Mmusculus))
anchor.two = GRanges(c("chr1", "chr1", "chr1", "chr2"), IRanges(c(100, 200, 300, 50), width=5), seqlengths=seqlengths(Mmusculus))
test <- new("GenomicInteractions", experiment_name="test", description="this is a test",
                 genome_name="BSgenome.Mmusculus.UCSC.mm9", anchor_one = anchor.one,
                 anchor_two = anchor.two, counts=as.integer(c(2,1,2,3)), pvalue=c(0.1, 0.3, 0.1, 0.08))
}

