% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{findOverlaps}
\alias{countOverlaps,GRanges,GenomicInteractions-method}
\alias{countOverlaps,GenomicInteractions,GRanges-method}
\alias{findOverlaps}
\alias{findOverlaps,GRanges,GenomicInteractions-method}
\alias{findOverlaps,GenomicInteractions,GRanges-method}
\alias{findOverlaps,GenomicInteractions,GenomicInteractions-method}
\alias{overlapsAny,GRanges,GenomicInteractions-method}
\alias{overlapsAny,GenomicInteractions,GRanges-method}
\title{Find overlaps between GRanges and GenomicInteractions objects}
\usage{
\S4method{findOverlaps}{GenomicInteractions,GRanges}(query, subject,
  maxgap = 0L, minoverlap = 1L, type = c("any", "start", "end", "within",
  "equal"), select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE)

\S4method{findOverlaps}{GRanges,GenomicInteractions}(query, subject,
  maxgap = 0L, minoverlap = 1L, type = c("any", "start", "end", "within",
  "equal"), select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE)

\S4method{countOverlaps}{GenomicInteractions,GRanges}(query, subject,
  maxgap = 0L, minoverlap = 1L, select = c("all", "first", "last",
  "arbitrary"), type = c("any", "start", "end", "within", "equal"),
  ignore.strand = FALSE)

\S4method{countOverlaps}{GRanges,GenomicInteractions}(query, subject,
  maxgap = 0L, minoverlap = 1L, type = c("any", "start", "end", "within",
  "equal"), select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE)

\S4method{overlapsAny}{GenomicInteractions,GRanges}(query, subject,
  maxgap = 0L, minoverlap = 1L, type = c("any", "start", "end", "within",
  "equal"), ignore.strand = FALSE)

\S4method{overlapsAny}{GRanges,GenomicInteractions}(query, subject,
  maxgap = 0L, minoverlap = 1L, type = c("any", "start", "end", "within",
  "equal"), ignore.strand = FALSE)

\S4method{findOverlaps}{GenomicInteractions,GenomicInteractions}(query, subject)
}
\arguments{
\item{query}{GenomicInteractions or GRanges}

\item{subject}{GRanges or GenomicInteractions}

\item{maxgap,minoverlap,type,select,ignore.strand}{See 'findOverlaps' in the IRanges package.}
}
\value{
A Hits object or a list containing Hits objects for both anchors.
}
\description{
When called with a GRanges and a GenomicInteractions object, this function
calls findOverlaps separately on each anchor and returns a list. countOverlaps
and overlapsAny return a list of integer vectors and logical vectors respectively.
}
\details{
When

See
'findOverlaps' in the GenomicRanges package for detailed
documentation for this function.
}

