% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-package.R
\docType{package}
\name{GenomicInteractions-package}
\alias{GenomicInteractions-package}
\title{GenomicInteractions}
\description{
R package for handling Genomic interaction data, such as
ChIA-PET/Hi-C, annotating genomic features with interaction
information and producing various plots / statistics.
}
\examples{
library(GenomicInteractions)
}
