% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[}
\alias{[}
\alias{[,GenomicInteractions,Rle,missing-method}
\alias{[,GenomicInteractions,logical,missing-method}
\alias{[,GenomicInteractions,numeric,missing-method}
\alias{[,GenomicInteractions,rle,missing-method}
\title{Standard subsetting methods for GenomicInteractions objects}
\usage{
\S4method{[}{GenomicInteractions,logical,missing}(x, i)

\S4method{[}{GenomicInteractions,numeric,missing}(x, i)

\S4method{[}{GenomicInteractions,Rle,missing}(x, i)

\S4method{[}{GenomicInteractions,rle,missing}(x, i)
}
\arguments{
\item{x}{A genomicInteractions object}

\item{i}{A numeric, logical or Rle vector}
}
\value{
A GenomicInteractions object containing only the features specified by `i`.
}
\description{
Standard subsetting methods for GenomicInteractions objects
}

