% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{subsetByFeatures}
\alias{subsetByFeatures}
\alias{subsetByFeatures,GenomicInteractions,GRanges,missing-method}
\alias{subsetByFeatures,GenomicInteractions,GRangesList,missing-method}
\alias{subsetByFeatures,GenomicInteractions,character,character-method}
\title{Subset a GenomicInteractions object by features}
\usage{
subsetByFeatures(GIObject, features, feature.class = NULL)

\S4method{subsetByFeatures}{GenomicInteractions,GRanges,missing}(GIObject,
  features, feature.class = NULL)

\S4method{subsetByFeatures}{GenomicInteractions,GRangesList,missing}(GIObject,
  features, feature.class = NULL)

\S4method{subsetByFeatures}{GenomicInteractions,character,character}(GIObject,
  features, feature.class = NULL)
}
\arguments{
\item{GIObject}{A GenomicInteractions object}

\item{features}{A GRanges or GRangesList object, or a character vector containing
IDs of annotated features, e.g. promoter IDs.}

\item{feature.class}{If `features' is a character vector, the corresponding feature name, e.g. "promoter".}
}
\value{
a subsetted GenomicInteractions object
}
\description{
Subsets interactions for which at least one of the anchors overlaps with a given GRanges object.
Alternatively, subsets interactions based on annotated feature IDs for a particular feature.
}

