% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-class.R
\name{GenomicInteractions}
\alias{GenomicInteractions}
\alias{GenomicInteractions,GRangesFactor,GRangesFactor,missing-method}
\alias{GenomicInteractions,GRanges,GRanges,missing-method}
\alias{GenomicInteractions,GRanges,GRanges,GRanges-method}
\alias{GenomicInteractions,integer,integer,GRanges-method}
\alias{GenomicInteractions-class}
\alias{anchors}
\alias{anchors,GenomicInteractions-method}
\alias{anchors<-}
\alias{anchors<-,GenomicInteractions-method}
\alias{regions}
\alias{regions,GenomicInteractions-method}
\alias{regions<-}
\alias{regions<-,GenomicInteractions-method}
\alias{first<-,GenomicInteractions-method}
\alias{second<-,GenomicInteractions-method}
\alias{$,GenomicInteractions-method}
\alias{$<-,GenomicInteractions-method}
\alias{show,GenomicInteractions-method}
\title{The GenomicInteractions class}
\description{
The GenomicInteractions class is derived from the \linkS4class{Pairs} class
and is dedicated to representing interactions between pairs of genomic intervals.
Each genomic interval is referred to as an \dQuote{anchor}, and each interaction is defined between a pair of anchors.
We will refer to the two anchors in a pair as the \dQuote{first} and \dQuote{second} anchors (arbitrarily defined).

All first anchors in a GenomicInteractions instance are represented by one \linkS4class{GRangesFactor} object,
while all second anchors are represented by another GRangesFactor object of the same length.
This improves efficiency by avoiding storage and processing of redundant copies of the same genomic coordinates and metadata;
especially for interaction data, where one interval may be involved in multiple interactions.

It is worth clarifying the terminology at this point.
The \dQuote{anchors} refer to the genomic intervals themselves;
the \dQuote{anchor type} refers to the anchor ordering, i.e., first or second;
the \dQuote{anchor levels} are the universe of possible anchors across all interactions for a given anchor type;
and the \dQuote{anchor indices} are integer indices to the anchor levels,
such that subsetting the anchor levels by the anchor indices yields the anchors.
}
\section{Constructors}{

\code{GenomicInteractions(anchor1, anchor2, regions, ..., metadata=list())}
will create a GenomicInteractions object, given:
\enumerate{
\item A \linkS4class{GRangesFactor} object in each of \code{anchor1} and \code{anchor2}, and missing \code{regions}.
\item A \linkS4class{GRanges} object in each of \code{anchor1} and \code{anchor2}, and missing \code{regions}.
\item A \linkS4class{GRanges} object in each of \code{anchor1} and \code{anchor2}, 
and a GRanges in \code{regions} that contains a superset of intervals in \code{anchor1} and \code{anchor2}.
\item An integer vector in each of \code{anchor1} and \code{anchor2}, and a GRanges in \code{regions}.
Anchor regions are defined by using integers to index \code{regions}.
}

Any arguments in \code{...} should be vector-like and are added to the element-wise metadata of the output object.
For options 1-3, any metadata in \code{anchor1} or \code{anchor2} are moved to the element-wise metadata,
with the anchor of origin prepended to the name.

Option 2 has an additional \code{common=TRUE} argument,
where the anchor levels for both the first and second anchors are defined as the union of the regions used in both anchors.
This is provided for backwards compatibility with the old GenomicInteractions behaviour,
and provides identical \code{regions} regardless of \code{type} (see below).

The \code{metadata} argument is expected to be a list-like object,
to be used as the \code{\link{metadata}} slot of the output object.
}

\section{Getters}{

In the following code snippets, \code{x} is a GenomicInteractions object:
\describe{
\item{\code{anchors(x, type="both", id=FALSE)}:}{Returns the anchor regions.
If \code{type="both"}, it returns a \linkS4class{Pairs} of length 2,
where the first and second entries contain a GRangesFactor of the first and second anchors, respectively.

A GRangesFactor of only the first anchor regions can be returned directly with \code{type=1} or \code{type="first"}.
The same applies for the second anchors with \code{type=2} or \code{type="second"}.

\code{id=TRUE} will return the integer indices pointing to the anchor regions in \code{regions(x)}.
If \code{type="both"}, this will be a \linkS4class{DataFrame} of integer vectors, with one column per partner.
Otherwise, the appropriate integer vector is directly returned for the partner specified by \code{type}.
}
\item{\code{regions(x, type=NA)}:}{Returns the universe of all regions corresponding to anchor region \code{type}.
Setting \code{type=1} or \code{"first"} returns a GRanges of all first anchors,
while setting \code{type=2} or \code{"second"} does the same for all second anchors.
Setting \code{type="both"} returns a List of length 2 containing the regions used in the first and second anchors.

\code{type=NA} behaves like \code{type=1} with a deprecation warning.
This is intended to encourage users to actually specify a value of \code{type} in calls to \code{regions(x)},
as this option was not available in previous versions.

Note that \code{type=1} and \code{type=2} may not yield the same result!
Anchor indices are only directly comparable if the universe of regions is the same for both anchors.
}
\item{\code{first(x)}:}{A synonym for \code{anchors(x, 1)}.}
\item{\code{second(x)}:}{A synonym for \code{anchors(x, 2)}.}
\item{\code{x$name}:}{Retrieves the column named \code{name} from \code{mcols(x)},
yielding some vector-like object with the same length as \code{x}.}
}
All getter methods applicable to \linkS4class{Vector} objects can also be used, 
e.g., \code{\link{mcols}}, \code{\link{metadata}}.
}

\section{Setters}{

In the following code snippets, \code{x} is a GenomicInteractions object:
\describe{
\item{\code{anchors(x, type="both", id=FALSE) <- value}:}{Replaces the anchor regions specified by \code{type} with \code{value}.
If \code{type="both"}, \code{value} should be a Pairs of two GRanges,
where the first and second entries are to be used as the replacement first and second anchor regions, respectively.
If \code{type=1} or \code{"first"}, \code{value} should be a GRanges or GRangesFactor to replace the first anchors;
same for \code{type=2} or \code{"second"} for the second anchors.

\code{id=TRUE} will replace the integer indices pointing to the anchor regions in \code{regions(x)}.
If \code{type="both"}, \code{value} should be a \linkS4class{DataFrame} of integer vectors, with one column per partner.
Otherwise, it should be an integer vector to replace the indices for the partner specified by \code{type}.
}
\item{\code{regions(x, type=NA) <- value}:}{Replaces the anchor levels in \code{x}.
If \code{type=1} or \code{"first"}, \code{value} should be a GRanges to replace the first anchor levels;
same for \code{type=2} or \code{"second"} for the second anchor levels.
If \code{type="both"}, \code{value} should be a List of length 2 to replace both levels simultaneously.

\code{type=NA} is provided for backwards compatibility, 
and will replace the region sets for \emph{both} anchors with the GRanges \code{value}.
This will trigger a deprecation warning - 
users wanting this behaviour should use \code{type="both"} with the replacement value set to \code{List(value, value)}.
}
\item{\code{first(x) <- value}:}{A synonym for \code{anchors(x, 1) <- value}.}
\item{\code{second(x) <- value}:}{A synonym for \code{anchors(x, 2) <- value}.}
\item{\code{x$name <- value}:}{Sets the column named \code{name} in \code{mcols(x)} with \code{value},
where \code{value} is recycled to have the same length as \code{x}.}
}
Again, all setter methods applicable to Vector objects can also be used here, 
e.g., \code{\link{mcols<-}}, \code{\link{metadata<-}}.
}

\section{Comparison methods}{

Methods like \code{sort}, \code{\link{match}} and \code{\link{pcompare}}
are supported via inheritance from the \linkS4class{Pairs} class.

One entry X of a GenomicInteractions object is considered \dQuote{greater} than another entry Y if the first anchor of X is \dQuote{greater} than that of Y; 
if these are equal, then if the second anchor of X is greater than the the second anchor of Y.
Comparisons between anchors are defined according to comparisons between \linkS4class{GRanges}, see \code{?"\link{GenomicRanges-comparison}"}.
}

\examples{
anchor1 <- GRanges(c("chr1", "chr1", "chr1", "chr1"), 
    IRanges(c(10, 20, 30, 20), width=5))
anchor1 <- sample(anchor1, 20, replace=TRUE)

anchor2 <- GRanges(c("chr1", "chr1", "chr1", "chr2"), 
    IRanges(c(100, 200, 300, 50), width=5))
anchor2 <- sample(anchor2, 20, replace=TRUE)

test <- GenomicInteractions(anchor1, anchor2)
test

# Alternative methods of construction:
combined <- sort(unique(c(anchor1, anchor2)))
m1 <- match(anchor1, combined)
m2 <- match(anchor1, combined)
test2 <- GenomicInteractions(m1, m2, combined)
test2

# Getting
anchors(test, 1)
anchors(test, 2) 
regions(test, 1)
regions(test, 2)

# Setting
anchors(test, 1) <- rev(anchors(test, 1))
test

first.regions <- regions(test, type=1)
regions(test, type=1) <- resize(first.regions, 1000)
test

}
\seealso{
\linkS4class{GRangesFactor}, used to represent the anchor regions.

\linkS4class{Vector}, from which this class is derived.
}
\author{
Aaron Lun
}
