% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{length,GenomicInteractions-method}
\alias{length,GenomicInteractions-method}
\title{Get the length of a GenomicInteractions GIObject}
\usage{
\S4method{length}{GenomicInteractions}(x)
}
\arguments{
\item{x}{GenomicInteractions GIObject}
}
\value{
A numeric vector containing the length of the GIObject
}
\description{
Get the length of a GenomicInteractions GIObject
}

