% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mm9_refseq_promoters}
\alias{mm9_refseq_promoters}
\title{Mouse Refseq promoters from chr 14-15}
\format{A GRanges object with length 2441.}
\usage{
data(mm9_refseq_promoters)
}
\description{
This dataset contains a subset of the promoters from the Refseq annotation
for mouse genome build mm9. See the HiC analysis vignette
(\code{vignettes(GenomicInteractions)}) for more information on how this
dataset was created.
}
\keyword{datasets}

