% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairdist.R
\docType{methods}
\name{pairdist}
\alias{pairdist}
\alias{intrachr}
\title{Get the linear distance for each interaction}
\usage{
\S4method{pairdist}{GenomicInteractions}(x, type = "mid")

intrachr(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicInteractions} object.}

\item{type}{String pecifying the type of distance to compute.
Can take values of \code{"mid"}, \code{"gap"}, \code{"span"}, \code{"diag"} or \code{"intra"}.}

\item{An}{integer vector of base-pair distances for \code{type="gap"} or \code{"span"}.

An integer vector of \dQuote{unit-based} distances for \code{type="diag"}.

A numeric vector of base-pair distances for \code{type="mid"}.

A logical vector for \code{type="intra"} or from \code{intrachr}.}
}
\description{
Compute the distance between interacting regions on the linear genome, for each pairwise interaction contained in a \linkS4class{GenomicInteractions} object.
}
\details{
For each interaction in \code{x}, the \code{pairdist} method computes the distance between the two interacting regions.
An integer vector is returned, with values computed according to the specified value of \code{type}:
\describe{
\item{\code{"mid"}:}{The distance between the midpoints of the two regions (rounded down to the nearest integer) is returned.}
\item{\code{"gap"}:}{The length of the gap between the closest points of the two regions is computed - 
negative lengths are returned for overlapping regions, indicating the length of the overlap.}
\item{\code{"span"}:}{The distance between the furthermost points of the two regions is computed.}
\item{\code{"diag"}:}{The difference between the anchor indices is returned.
This corresponds to a diagonal on the interaction space when the regions are genomic bins,
and requires that both anchors have the same universe of \code{\link{regions}}
(an error is raised otherwise).}
}

Interchromosomal interactions are marked with \code{NA}.
Alternatively, if \code{type="intra"}, a logical vector is returned indicating whether the interaction occurs between two regions on the same chromosome.
\code{intrachr(x)} is an alias for \code{pairdist(x, type="intra")}.
}
\examples{
anchor1 <- GRanges(c("chr1", "chr1", "chr1", "chr1"), 
    IRanges(c(10, 20, 30, 20), width=5))
anchor2 <- GRanges(c("chr1", "chr1", "chr1", "chr2"), 
    IRanges(c(100, 200, 300, 50), width=5))
test <- GenomicInteractions(anchor1, anchor2)
test

pairdist(test)
pairdist(test, type="gap")
pairdist(test, type="span")
intrachr(test)

# Common regions.
regions <- GRanges("chr1", IRanges(sort(sample(1000, 10)), width=5))
i1 <- sample(length(regions), 10)
i2 <- sample(length(regions), 10)
test.common <- GenomicInteractions(i1, i2, regions)
pairdist(test.common, type="diag")

}
\author{
Aaron Lun
}
