% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummaryStats.R
\name{plotCisTrans}
\alias{plotCisTrans}
\title{Plots the percentages of cis and trans interactions for a GInteractions object as a donut plot.}
\usage{
plotCisTrans(GIObject)
}
\arguments{
\item{GIObject}{A GInteractions object}
}
\value{
A ggplot2 plot
}
\description{
Plots the percentages of cis and trans interactions for a GInteractions object as a donut plot.
}
\examples{
data(hic_example_data)
plotCisTrans(hic_example_data)
}

