% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSummaryStats.R
\name{plotDists}
\alias{plotDists}
\title{Plots a histogram of interaction distances for a GenomicInteractions Object}
\usage{
plotDists(GIObject, breaks = c(0, 1000, 5000, 10000, 50000, 1e+05, 5e+05,
  1e+06, 2e+06), method = "midpoint")
}
\arguments{
\item{GIObject}{A GenomicInteractions object}

\item{breaks}{A numeric vector of breaks for the histogram}

\item{method}{Method used for distance between anchors. Passed to calculateDistances. One of "midpoint", "inner", or "outer".}
}
\value{
A ggplot2 plot
}
\description{
Plots a histogram of interaction distances for a GenomicInteractions Object
}
\examples{
data(hic_example_data)
plotDists(hic_example_data)
}

