% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotRegion.R
\docType{methods}
\name{plotRegion}
\alias{plotRegion}
\alias{plotRegion,GenomicInteractions,GRanges,list-method}
\title{Plot interactions within a specified region.}
\usage{
plotRegion(GIObject, region, annotation.features, annotation.cols = NULL,
  reduce.anchors = TRUE, plot.trans = TRUE, plot.cis = TRUE,
  order.cis = TRUE, plot.cis.names = TRUE, plot.header = TRUE,
  plot.lines = TRUE, anchor.col = "darkred", plot.ids = FALSE)

\S4method{plotRegion}{GenomicInteractions,GRanges,list}(GIObject, region,
  annotation.features, annotation.cols = NULL, reduce.anchors = TRUE,
  plot.trans = TRUE, plot.cis = TRUE, order.cis = TRUE,
  plot.cis.names = TRUE, plot.header = TRUE, plot.lines = TRUE,
  anchor.col = "darkred", plot.ids = FALSE)
}
\arguments{
\item{GIObject}{GenomicInteractions object}

\item{region}{A GRanges specifying the genomic region to plot}

\item{annotation.features}{a list of GRanges specifying the features within the region to plot}

\item{annotation.cols}{a named vector specifying which colour to plot the individual tracks}

\item{reduce.anchors}{a logical specifying whether to reduce the anchor GRanges}

\item{plot.trans}{a logical specifying whether to show trans-interactions}

\item{plot.cis}{a logical specifying whether to plot cis-interactions that are outside of the specified region}

\item{order.cis}{logical specifying whether to order cis-interactions by their distances to the specified region}

\item{plot.cis.names}{a logical specifying whether to plot textual information on the other anchor region of cis-interactions}

\item{plot.header}{a logical specifying whether to plot a header describing the genomic region plotted}

\item{plot.lines}{a logical specifying whether to plot dashed lines indicating anchor regions and associated features}

\item{anchor.col}{colour for anchor regions}

\item{plot.ids}{a logical specifying whether to plot ids for features in annotation.features. Looks for the presence of an id or name}
}
\value{
invisible(1)
}
\description{
This is function allows the plotting of an interactions between annotated features in a specified area.
The resulting plot shows unique interactions as curves between interaction anchor points with the number
of counts supporting that interaction proportional to the thickness of that line. It is also possible to
add cis-interactions which are not within the window/region and to also plot regions that are involved
in trans-interactions. Plotting the data this way makes it possible to examine a region and easily examine
which regions are highly interacting with each other. It is not recommended to use this style of plot to examine
regions larger than 5Mb.
}
\examples{
\dontrun{
k562.rep1 = GenomicInteractions(file.path(system.file(package="GenomicInteractions"), "extdata", "k562.rep1.cluster.pet3+.txt"),
                                     type="chiapet.tool",
                                     experiment_name="k562",
                                     description="k562 pol2 8wg16",
                                     gname="BSgenome.Hsapiens.UCSC.hg19")
annotation.features = list(promoter=refseq.promoters,
                             terminator=refseq.terminators,
                             gene.body=refseq.transcripts)
plotRegion(k562.rep1, GRanges("chr18", IRanges(9000000, 10000000)), annotation.features)
}
}

