% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/viewPoints.R
\name{plotViewpoint}
\alias{plotViewpoint}
\title{Plot coverage around a virtual 4C viewpoint}
\usage{
plotViewpoint(x, region, ...)
}
\arguments{
\item{x}{a GenomicInteractions object which is output from viewPoint}

\item{region}{The genomic region to plot}

\item{...}{additional arguments to plot}
}
\value{
output of plot()
}
\description{
Plot coverage around a virtual 4C viewpoint
}

