% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{print,GenomicInteractions-method}
\alias{print,GenomicInteractions-method}
\title{Print function for GenomicInteractions}
\usage{
\S4method{print}{GenomicInteractions}(x)
}
\arguments{
\item{x}{GenomicInteractionsObject}
}
\value{
invisible(1)
}
\description{
Print function for GenomicInteractions
}

