% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sameStrand}
\alias{sameStrand}
\title{Tests whether anchors have the same strand.}
\usage{
sameStrand(GIObject)
}
\arguments{
\item{GIObject}{A GenomicInteractions object}
}
\value{
A logical vector denoting with TRUE if both anchors of an interaction
 are on the same strand and FALSE otherwise.
}
\description{
This is designed for processing .bam files.
}

