% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{seqinfo,GenomicInteractions-method}
\alias{seqinfo,GenomicInteractions-method}
\alias{seqinfo<-,GenomicInteractions-method}
\title{Acessing/modifying sequence information for a GenomicInteractions object}
\usage{
\S4method{seqinfo}{GenomicInteractions}(x)

\S4method{seqinfo}{GenomicInteractions}(x, new2old = NULL, force = FALSE) <-
  value
}
\arguments{
\item{x}{A GenomicInteractions object}

\item{new2old}{Mapping between new and old seqnames. See ?seqinfo in GenomeInfoDb for details.}

\item{force}{A logical indicating whether or not to drop invalid levels.}

\item{value}{A replacement seqinfo object}
}
\value{
A seqinfo object,
}
\description{
Allows access/modification of seqinfo for GenomicInteractions objcets. When
used with "force=True", interactions with either (or both) anchors on invalid
chromosomes will be removed.
}
\details{
For more information see ?seqinfo in the GenomeInfoDb
package.
}

