% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sort,GenomicInteractions-method}
\alias{sort,GenomicInteractions-method}
\title{Sort GenomicInteractions Object}
\usage{
\S4method{sort}{GenomicInteractions}(x, decreasing = FALSE,
  order.interactions = TRUE)
}
\arguments{
\item{x}{GenomicInteractions Object}

\item{decreasing}{A logical indicating sort order}

\item{order.interactions}{A logical indicating if interactions should be reordered,
or only rearranged so that start(anchorOne) < start(anchorTwo)}
}
\value{
A sorted GenomicInteractions object
}
\description{
This method will sort a GenomicInteractions object by first
arranging all interactions start on the lower-ordered anchor;
for trans-chromosomal interactions this is the anchor on the
lower ordered chromomsome (defined by the seqlevels factor);
and then by the position of the first anchor. See
?"GenomicRanges-comparison" for ordering rules.
}

