% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{sum,GInteractions-method}
\alias{sum,GInteractions-method}
\title{Return the total number of interactions in a GInteractions GIObject}
\usage{
\S4method{sum}{GInteractions}(x)
}
\arguments{
\item{x}{GInteractions GIObject}
}
\value{
The sum of the counts in GIObject
}
\description{
Return the total number of interactions in a GInteractions GIObject
}

