% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotators.R
\docType{methods}
\name{summariseByFeaturePairs}
\alias{summariseByFeaturePairs}
\alias{summariseByFeaturePairs,GenomicInteractions-method}
\title{Summarise the number of interactions between two sets of features.}
\usage{
summariseByFeaturePairs(GIObject, features.one, feature.name.one, features.two,
  feature.name.two)

\S4method{summariseByFeaturePairs}{GenomicInteractions}(GIObject, features.one,
  feature.name.one, features.two, feature.name.two)
}
\arguments{
\item{GIObject}{An annotated GenomicInteractions object}

\item{features.one}{A GRanges object containing the feature set of interest}

\item{feature.name.one}{The name of the first feature set of interest}

\item{features.two}{A GRanges object containing the second feature set of interest}

\item{feature.name.two}{The name of the second feature set of interest}
}
\value{
A data frame with one line for each range in `features'
}
\description{
This function will calculate the number of observed interactions between
two sets of features provided by the end-user. This allows the summarisation
of the number of features of a specific type a particular region is involved in
and how many interactions exist between them.
}

