% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapAnchors.R
\docType{methods}
\name{swapAnchors}
\alias{swapAnchors}
\alias{swapAnchors,GenomicInteractions-method}
\title{Swap anchor regions}
\usage{
\S4method{swapAnchors}{GenomicInteractions}(x, mode = c("order",
  "reverse", "all"))
}
\arguments{
\item{x}{A \linkS4class{GenomicInteractions} object.}

\item{mode}{String specifying the type of swapping to perform.}
}
\value{
A GenomicInteractions object with some or all of first and second anchor regions swapped.
}
\description{
Swap anchor regions within an interaction to ensure that the first anchor is not greater than the second.
}
\details{
\code{mode="order"} will ensure that the first anchor region is not ordered above the second anchor in the output object.

\code{mode="reverse"} will ensure that the first anchor region is not ordered below the second anchor in the output.

\code{mode="all"} will simply swap the first and second anchor regions without any consideration of ordering.
}
\examples{
anchor1 <- GRanges(c("chr1", "chr1", "chr2", "chr1"), 
    IRanges(c(10, 20, 30, 20), width=5))
anchor2 <- GRanges(c("chr1", "chr1", "chr1", "chr2"), 
    IRanges(c(5, 50, 3, 50), width=5))
test <- GenomicInteractions(anchor1, anchor2)
test

swapAnchors(test)
swapAnchors(test, mode="reverse")
swapAnchors(test, mode="all")

}
\author{
Aaron Lun
}
