% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentData.R
\docType{data}
\name{thymus_enh}
\alias{thymus_enh}
\title{Putative enhancers from mouse thymus data}
\format{A GRanges object}
\usage{
data("thymus_enhancers")
}
\value{
A GRanges object
}
\description{
This dataset contains a set of mouse thymus enhancers derived from ChIP-seq data
from mouse thymus, as described in Shen et al. 2012. See the HiC analysis vignette 
for more details. (\code{vignettes(GenomicInteractions)}).
}
\references{
Shen, Y et al. A map of cis-regulatory sequences in the mouse genome. Nature (2012).
}
\keyword{datasets}

