% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{trim,GenomicInteractions-method}
\alias{trim,GenomicInteractions-method}
\title{Trim a GenomicInteractions object}
\usage{
\S4method{trim}{GenomicInteractions}(x, minAnchorSize = 1, ...)
}
\arguments{
\item{x}{A GenomicInteractions object}

\item{minAnchorSize}{The minimum size anchor to allow when trimming ranges.}

\item{...}{any additional arguments to trim}
}
\value{
A trimmed GenomicInteractions object
}
\description{
This will remove any interactions with an anchor falling outside
of the seqlengths in a GenomicInteractions object, and trim ranges
which cross the ends of chromosomes.
}

