% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateObject.R
\docType{methods}
\name{updateObject,GenomicInteractions-method}
\alias{updateObject,GenomicInteractions-method}
\title{Update old GenomicInteractions}
\usage{
\S4method{updateObject}{GenomicInteractions}(object, ...,
  verbose = FALSE)
}
\arguments{
\item{object}{Any instance of an old GenomicInteractions class, 
generated prior to version 2.0.0 of the \pkg{GenomicInteractions} package.}

\item{...}{Further arguments, not used.}

\item{verbose}{Logical scalar indicating whether to display messages during conversion.}
}
\value{
A \linkS4class{GenomicInteractions0} object.
}
\description{
Update an old GenomicInteractions instance (usually saved to file in serialized form)
into a backwards-compatible \linkS4class{GenomicInteractions0} object.
}
\details{
We return a \linkS4class{GenomicInteractions0} instance rather than a \linkS4class{GenomicInteractions} instance,
simply to maximize backwards-compatibility with code that calls \code{\link{regions}} 
and expects a \linkS4class{GenomicRanges} in return.
}
\examples{
\dontrun{
updateObject(readRDS('some_old_objects.rds'))
}

}
\author{
Aaron Lun,
based on code by Malcolm Perry and Liz Ing-Simmons.
}
