% Generated by roxygen2 (4.0.2): do not edit by hand
\name{viewPoint}
\alias{viewPoint}
\title{Plot coverage of interactions originating at a given viewpoint.}
\usage{
viewPoint(pos, GIObject, leftflank, rightflank, plot = TRUE)
}
\arguments{
\item{pos}{A single region in GRanges format.}

\item{GIObject}{A GenomicInteractions object.}

\item{leftflank}{An integer; flank size in bp upstream of pos centre.}

\item{rightflank}{An integer; flank size in bp downstream of pos centre.}

\item{plot}{Logical. Whether to plot vector (default), or just return vector of coverage.}
}
\value{
Plot of coverage or Rle-vector of coverage, depending on plot parameter.
}
\description{
Plot coverage of interactions originating at a given viewpoint.
}
\examples{
\dontrun{
data(hic_data)
pos <- GRanges(seqnames="chr5", ranges=IRanges(start=115938063, end=115941352))
viewPoint(pos,hic_data, 100000, 100000, plot=FALSE)
}
}

