% Generated by roxygen2 (4.0.2): do not edit by hand
\name{viewPointAverage}
\alias{viewPointAverage}
\title{Plot coverage of interactions originating at a given set of viewpoints.}
\usage{
viewPointAverage(positions, GIObject, leftflank, rightflank, plot = TRUE,
  flip = FALSE)
}
\arguments{
\item{positions}{A set of regions in GRanges format.}

\item{GIObject}{A GenomicInteractions object.}

\item{leftflank}{An integer; flank size in bp upstream of position centre.}

\item{rightflank}{An integer; flank size in bp downstream of position centre.}

\item{plot}{Logical. Whether to plot vector (default), or just return vector of coverage.}

\item{flip}{Logical. If TRUE, flip window of coverage around a position on the minus strand. Use e.g. to plot around promoters. Default: FALSE.}
}
\value{
Plot of mean coverage around all positions or Rle-vector of mean coverage, depending on plot parameter.
}
\description{
Plot coverage of interactions originating at a given set of viewpoints.
}

