% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotators.R
\docType{methods}
\name{annotateInteractions}
\alias{annotateInteractions}
\alias{annotateInteractions,GenomicInteractions,list-method}
\title{Annotate the interactions in a GenomicInteractions object}
\usage{
annotateInteractions(GIObject, annotations)

\S4method{annotateInteractions}{GenomicInteractions,list}(GIObject, annotations)
}
\arguments{
\item{GIObject}{A GenomicInteractions object to be annotated}

\item{annotations}{A list containing GRanges (or GRangesList) objects with which to annotate
the GenomicInteractions object.}
}
\value{
invisible(1)
}
\description{
This function will annotate both anchors with a list of named GRanges
objects. Each metadata column is labeled "name.id" and contains the id of
the genomic interval(s) it overlaps. Anonymous lists will be given names
"FEATURE#.id" where # is the position in the list.
}
\details{
For each anchor a "node.class" metadata column will also be added, containing
the name of the list element which was \emph{first} annotated to each range.
Ranges with no overlaps will be classified as "distal". The identifiers for each
individual feature/annotation are taken from either the name of the list item in the
case of a GRangesList or from either the names of a the provided GRanges or an id column
in its associated metadata.
}
\examples{
data(hic_example_data)
data(mm9_refseq_promoters)
\dontrun{
annotateInteractions(hic_example_data, list(promoter=mm9_refseq_promoters))
}
}

