% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSummaryStats.R
\name{categoriseInteractions}
\alias{categoriseInteractions}
\title{Get the numbers of interaction types existing in your data}
\usage{
categoriseInteractions(GIObject, node.classes = NULL, viewpoints = NULL)
}
\arguments{
\item{GIObject}{A GenomicInteractions object}

\item{node.classes}{Optional. All node.classes to include in the analysis.
Default: all node classes.}

\item{viewpoints}{Optional. If set will only consider interactions where at
least one anchor is of this node class. Default: all classes in node.classes.}
}
\value{
A data.frame.
}
\description{
Get the numbers of interaction types existing in your data
}
\examples{
data(hic_example_data)
categoriseInteractions(hic_example_data)
}

