% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{export.bedpe}
\alias{export.bedpe}
\alias{export.bedpe,GenomicInteractions-method}
\title{Export interactions in BED Paired-End format.}
\usage{
export.bedpe(GIObject, fn = NULL, score = "counts")

\S4method{export.bedpe}{GenomicInteractions}(GIObject, fn = NULL,
  score = "counts")
}
\arguments{
\item{GIObject}{A GenomicInteractions object.}

\item{fn}{A filename to write the interactions data to}

\item{score}{Which metadata column to use as the score: counts, pvalue, fdr, normalised}
}
\value{
invisible(1) if outputting to file or a data.frame containing all of the corresponding information
}
\description{
#' Exports a GenomicInteractions object to BED-PE format, and writes to a specified file. If filename is not specified,
then a data.frame containing the information is returned. The value of the score parameter defines which field is used
to populate the score field.
}

