% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GenomicInteractions-class.R
\docType{methods}
\name{$,GenomicInteractions-method}
\alias{$,GenomicInteractions-method}
\alias{$<-,GenomicInteractions-method}
\title{Quick access to GenomicInteractions metadata columns}
\usage{
\S4method{$}{GenomicInteractions}(x, name)

\S4method{$}{GenomicInteractions}(x, name) <- value
}
\arguments{
\item{x}{A GenomicInteractions object}

\item{name}{Column to select}

\item{value}{Replacement value}
}
\value{
A vector containing the contents of the column
}
\description{
Quick access to GenomicInteractions metadata columns
}

