% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{findOverlaps}
\alias{findOverlaps}
\alias{findOverlaps,GRanges,GenomicInteractions-method}
\alias{findOverlaps,GenomicInteractions,GRanges-method}
\title{Find overlaps between a GRanges and a GenomicInteractions object}
\usage{
\S4method{findOverlaps}{GenomicInteractions,GRanges}(query, subject,
  maxgap = 0L, minoverlap = 1L, type = c("any", "start", "end", "within",
  "equal"), select = c("all", "first", "last", "arbitrary"),
  algorithm = c("nclist", "intervaltree"))

\S4method{findOverlaps}{GRanges,GenomicInteractions}(query, subject,
  maxgap = 0L, minoverlap = 1L, type = c("any", "start", "end", "within",
  "equal"), select = c("all", "first", "last", "arbitrary"),
  algorithm = c("nclist", "intervaltree"))
}
\arguments{
\item{query}{GenomicInteractions or GRanges}

\item{subject}{GRanges or GenomicInteractions}

\item{maxgap,minoverlap,type,select,algorithm}{See 'findOverlaps' in the IRanges package.}
}
\value{
A list containing Hits objects for anchors one and two.
}
\description{
This function calls findOverlaps separately on each anchor and
returns a list. See 'findOverlaps' in the GenomicRanges package for detailed
documentation for this function.
}

