% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-package.R
\docType{package}
\name{GenomicInteractions-package}
\alias{GenomicInteractions-package}
\title{GenomicInteractions}
\description{
Utilities for handling genomic interaction data such as
ChIA-PET or Hi-C, annotating genomic features with interaction
information, and producing plots and summary statistics.
}
\examples{
library(GenomicInteractions)
}
