% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GenomicInteractions-package.R
\docType{package}
\name{GenomicInteractions-package}
\alias{GenomicInteractions-package}
\title{A package for looking at genomic interaction data.}
\description{
A package for looking at genomic interaction data.
}

