% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-package.R
\docType{package}
\name{GenomicInteractions-package}
\alias{GenomicInteractions-package}
\title{GenomicInteractions: A package for analysing chromosome conformation}
\description{
GenomicInteractions: A package for analysing chromosome conformation
}
\examples{
library(GenomicInteractions)
}

