% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-package.R
\docType{package}
\name{GenomicInteractions-package}
\alias{GenomicInteractions}
\alias{GenomicInteractions-package}
\title{A package for looking at genomic interaction data.}
\description{
@name GenomicInteractions-package
 @docType package
 @import methods
 @examples 
 library(GenomicInteractions)
 GenomicInteractions()
}

