% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-package.R
\docType{package}
\name{GenomicInteractions-package}
\alias{GenomicInteractions-package}
\title{The \pkg{GenomicInteractions} package}
\description{
The \pkg{GenomicInteractions} package provides basic classes and methods 
for representing and manipulating chromosome conformation data.
This includes the \linkS4class{GenomicInteractions} class 
with methods for computing linear distances, finding the bounding box and identifying overlaps in one and two dimensions.
}
\examples{
library(GenomicInteractions)
}
