% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenomicInteractions-class.R
\name{GenomicInteractions}
\alias{GenomicInteractions}
\title{Function to create a GenomicInteraction object}
\usage{
GenomicInteractions(anchor_one = GRanges(), anchor_two = GRanges(),
  counts = integer(), experiment_name = NULL, description = NULL, ...)
}
\arguments{
\item{anchor_one,anchor_two}{GRanges objects.}

\item{counts}{An integer vector, defaults to 1.}

\item{experiment_name}{Experiment name.}

\item{description}{Description of experiment.}

\item{...}{Additional data to be added to mcols}
}
\value{
a GenomicInteractions object
}
\description{
Create GenomicInteraction objects from two GRanges ojects.
}
\examples{
library(GenomicRanges)

anchor.one = GRanges(c("chr1", "chr1", "chr1", "chr1"), IRanges(c(10, 20, 30, 20), width=5))
anchor.two = GRanges(c("chr1", "chr1", "chr1", "chr2"), IRanges(c(100, 200, 300, 50), width=5))
interaction_counts = sample(1:10, 4)
test <- GenomicInteractions(anchor.one, anchor.two, experiment_name="test",
                           description="this is a test", counts=interaction_counts)
}

