% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GenomicInteractions}
\alias{GenomicInteractions}
\title{Function to create GenomicInteraction objects.}
\usage{
GenomicInteractions(fn, type, experiment_name, description, gname)
}
\arguments{
\item{fn}{Filename or, if type="hiclib", folder}

\item{type}{One of "chiapet.tool", "chiapet.encode", "bed12", "bedpe", "hiclib", "homer", "bam".}

\item{experiment_name}{Experiment name.}

\item{description}{Description of experiment.}

\item{gname}{Genome name to use for constructing the GenomicInteractions object.}
}
\value{
a GenomicInteractions object
}
\description{
Function to create GenomicInteraction objects from a variety of files. The resulting objects contain information
on which genomic regions are interacting with each other, and the number of counts supporting each interaction.
It is also possible to store information on associated p-values and false-discovery rates (FDR).
It is possible to create GenomicInteractions objects for various datasets including Hi-C and ChIA-PET. It is possible
to read interactions from a variety of files including BAM files, bed files (BED12 and BEDPE) and from the output
from standard processing pipelines, such as HOMER and ChIA-PET tool. GenomicInteractions objects can also be created
using calls of the form \code{new("GenomicInteractions", ...)}. For hiclib, it expects the directory in which the files
extracted using h5dictToTxt.py from the hdf5 file are located, where as for all of the other file types it expects the full
filename.
}
\examples{
library(BSgenome.Hsapiens.UCSC.hg19)

k562.rep1 = GenomicInteractions(file.path(system.file(package="GenomicInteractions"), "extdata", "k562.rep1.cluster.pet3+.txt"),
                                     type="chiapet.tool",
                                     experiment_name="k562",
                                     description="k562 pol2 8wg16",
                                     gname="BSgenome.Hsapiens.UCSC.hg19")

k562.rep1
}

