% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicInteractions-class.R
\name{GenomicInteractions}
\alias{GenomicInteractions}
\alias{anchors}
\alias{anchors,GenomicInteractions-method}
\alias{anchors<-}
\alias{anchors<-,GenomicInteractions-method}
\alias{regions}
\alias{regions,GenomicInteractions-method}
\alias{regions<-}
\alias{regions<-,GenomicInteractions-method}
\alias{first,GenomicInteractions-method}
\alias{second,GenomicInteractions-method}
\alias{first<-,GenomicInteractions-method}
\alias{second<-,GenomicInteractions-method}
\title{The GenomicInteractions class}
\description{
The GenomicInteractions class is a subclass of the \linkS4class{IndexedRelations} class,
dedicated to representing interactions between pairs of genomic intervals.
}
\section{Overview}{

The main specializations from the base \linkS4class{IndexedRelations} class are:
\itemize{
\item The \code{featureSets} slot must be a \linkS4class{SimpleGenomicRangesList} object.
\item There must be exactly two partners for each relationship.
}

In short, each individual feature set is now a \linkS4class{GenomicRanges}.
Partners are referred to as \dQuote{anchor regions}, so called as they anchor the interactions between genomic loci.
}

\section{Constructors}{

\code{GenomicInteractions(anchor1, anchor2, regions, ...)} will create a GenomicInteractions object, given:
\itemize{
\item A \linkS4class{GenomicRanges} object in each of \code{anchor1} and \code{anchor2}, and missing \code{regions}.
\item An integer vector in each of \code{anchor1} and \code{anchor2}, 
and a \linkS4class{SimpleGenomicRangesList} of length 2 in \code{regions}.
\item An integer vector in each of \code{anchor1} and \code{anchor2}, and a GenomicRanges in \code{regions}.
}

Note that only the last option will yield a GenomicInteractions object with a single GenomicRanges as the feature set.
This is useful in situations where the first and second anchors have the same origin, e.g., genomic bins.
For the other options, the output object will contain a List of two GenomicRanges (one per anchor region).
This is useful in situations where the anchors are distinct, e.g., promoters versus enhancers.

Any arguments in \code{...} are added to the element-wise metadata of the output object.
}

\section{Getters}{

In the following code snippets, \code{x} is a GenomicInteractions object:
\describe{
\item{\code{anchors(x, type)}:}{Returns the anchor regions specified by \code{type} as a GenomicRanges object.
\code{type} should be an integer scalar specifying the first (\code{1}) or second anchor (\code{2}).}
\item{\code{regions(x)}:}{Returns the feature sets of all regions as a SimpleGenomicRangesList object.
The returned object can be of length 1 or 2, depending on how \code{x} was constructed.}
\item{\code{first(x)}:}{A synonym for \code{anchors(x, 1)}.}
\item{\code{second(x)}:}{A synonym for \code{anchors(x, 2)}.}
}
All getter methods applicable to IndexedRelations objects can also be used, e.g., \code{\link{partners}}, \code{\link{featureSets}}, \code{\link{mapping}}.
}

\section{Setters}{

In the following code snippets, \code{x} is a GenomicInteractions object:
\describe{
\item{\code{anchors(x, type) <- value}:}{Replaces the anchor regions specified by \code{type} with the entries of \code{value}, a GenomicRanges object.
\code{type} should be an integer scalar specifying the first (\code{1}) or second anchor (\code{2}).}
\item{\code{regions(x) <- value}:}{Replaces the feature sets of all regions as with \code{value}, a SimpleGenomicRangesList object.
\code{lengths(value)} should be of the same as \code{lengths(regions(x))} to ensure validity of the modified object.}
\item{\code{first(x) <- value}:}{A synonym for \code{anchors(x, 1) <- value}.}
\item{\code{second(x) <- value}:}{A synonym for \code{anchors(x, 2) <- value}.}
}
Again, all setter methods applicable to IndexedRelations objects can also be used, e.g., \code{\link{partners<-}}, \code{\link{featureSets<-}}.
}

\examples{
anchor1 <- GRanges(c("chr1", "chr1", "chr1", "chr1"), 
    IRanges(c(10, 20, 30, 20), width=5))
anchor2 <- GRanges(c("chr1", "chr1", "chr1", "chr2"), 
    IRanges(c(100, 200, 300, 50), width=5))
test <- GenomicInteractions(anchor1, anchor2)
test

# Getting
anchors(test, 1)
anchors(test, 2) 
regions(test)

# Setting
anchors(test, 1) <- rev(anchors(test, 1))
test
regions(test)[[1]] <- resize(regions(test)[[1]], 1000)
test

}
\seealso{
\linkS4class{IndexedRelations}, from which this class is derived.
}
\author{
Aaron Lun
}
