% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getters}
\alias{FDR}
\alias{FDR,GenomicInteractions-method}
\alias{anchorOne}
\alias{anchorOne,GenomicInteractions-method}
\alias{anchorTwo}
\alias{anchorTwo,GenomicInteractions-method}
\alias{annotationFeatures}
\alias{annotationFeatures,GenomicInteractions-method}
\alias{count}
\alias{count,GenomicInteractions-method}
\alias{description}
\alias{description,GenomicInteractions-method}
\alias{genomeName}
\alias{genomeName,GenomicInteractions-method}
\alias{getters}
\alias{name}
\alias{name,GenomicInteractions-method}
\alias{normalisedCount}
\alias{normalisedCount,GenomicInteractions-method}
\alias{pValue}
\alias{pValue,GenomicInteractions-method}
\title{Functions to access data held in a GenomicInteractions object.}
\usage{
name(GIObject)

anchorOne(GIObject)

anchorTwo(GIObject)

count(GIObject)

pValue(GIObject)

FDR(GIObject)

normalisedCount(GIObject)

description(GIObject)

genomeName(GIObject)

annotationFeatures(GIObject)

\S4method{name}{GenomicInteractions}(GIObject)

\S4method{anchorOne}{GenomicInteractions}(GIObject)

\S4method{anchorTwo}{GenomicInteractions}(GIObject)

\S4method{count}{GenomicInteractions}(GIObject)

\S4method{normalisedCount}{GenomicInteractions}(GIObject)

\S4method{pValue}{GenomicInteractions}(GIObject)

\S4method{FDR}{GenomicInteractions}(GIObject)

\S4method{description}{GenomicInteractions}(GIObject)

\S4method{genomeName}{GenomicInteractions}(GIObject)

\S4method{annotationFeatures}{GenomicInteractions}(GIObject)
}
\arguments{
\item{GIObject}{A GenomicInteractions object}
}
\value{
For 'anchorOne' and 'anchorTwo', a GRanges. For 'counts',
 'normalisedCount', pValue', 'FDR', a numeric vector with counts,
 normalised counts, p-values or FDRs for each interaction in the object. For
  'description','name', and 'genomeName', a character vector with
  length 1. For 'annotationFeatures', a character vector of features with
  which the object was previously annotated, or 'NA' if the object is unannotated.
}
\description{
Use these functions to access data stored in each of the slots of a
GenomicInteractions object.
}
\examples{
library(BSgenome.Mmusculus.UCSC.mm9)
anchor.one = GRanges(c("chr1", "chr1", "chr1", "chr1"), IRanges(c(10, 20, 30, 20), width=5), seqlengths=seqlengths(Mmusculus))
anchor.two = GRanges(c("chr1", "chr1", "chr1", "chr2"), IRanges(c(100, 200, 300, 50), width=5), seqlengths=seqlengths(Mmusculus))
test <- new("GenomicInteractions", experiment_name="test", description="this is a test",
                 genome_name="BSgenome.Mmusculus.UCSC.mm9", anchor_one = anchor.one,
                 anchor_two = anchor.two, counts=as.integer(c(2,1,2,3)), pvalue=c(0.1, 0.3, 0.1, 0.08))

name(test)
description(test)
anchorOne(test)
anchorTwo(test)
count(test)
pValue(test)
genomeName(test)
}

