% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getters.R
\docType{methods}
\name{getters}
\alias{anchorOne}
\alias{anchorOne,GenomicInteractions-method}
\alias{anchorTwo}
\alias{anchorTwo,GenomicInteractions-method}
\alias{annotationFeatures}
\alias{annotationFeatures,GenomicInteractions-method}
\alias{description}
\alias{description,GenomicInteractions-method}
\alias{getters}
\alias{interactionCounts}
\alias{interactionCounts,GenomicInteractions-method}
\alias{name}
\alias{name,GenomicInteractions-method}
\title{Functions to access data held in a GenomicInteractions object.}
\usage{
name(GIObject)

description(GIObject)

anchorOne(GIObject)

anchorTwo(GIObject)

interactionCounts(GIObject)

annotationFeatures(GIObject)

\S4method{name}{GenomicInteractions}(GIObject)

\S4method{description}{GenomicInteractions}(GIObject)

\S4method{anchorOne}{GenomicInteractions}(GIObject)

\S4method{anchorTwo}{GenomicInteractions}(GIObject)

\S4method{interactionCounts}{GenomicInteractions}(GIObject)

\S4method{annotationFeatures}{GenomicInteractions}(GIObject)
}
\arguments{
\item{GIObject}{A GenomicInteractions object}
}
\value{
For 'anchorOne' and 'anchorTwo', a GRanges. For 'interactionCounts',
a numeric vector with counts for each interaction in the object. For
  'description' and 'name',  a character vector with
  length 1. For 'annotationFeatures', a character vector of features with
  which the object was previously annotated, or 'NA' if the object is unannotated.
}
\description{
Use these functions to access data stored in each of the slots of a
GenomicInteractions object.
}
\examples{
library(GenomicRanges)

anchor.one = GRanges(c("chr1", "chr1", "chr1", "chr1"), IRanges(c(10, 20, 30, 20), width=5))
anchor.two = GRanges(c("chr1", "chr1", "chr1", "chr2"), IRanges(c(100, 200, 300, 50), width=5))
interaction_counts = sample(1:10, 4)
test <- GenomicInteractions(anchor.one, anchor.two, experiment_name="test",
                           description="this is a test", counts=interaction_counts)

name(test)
description(test)
anchorOne(test)
anchorTwo(test)
interactionCounts(test)
}

