% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DocumentData.R
\docType{data}
\name{hg19.refseq.transcripts}
\alias{hg19.refseq.transcripts}
\title{Human Refseq transcripts from chr 17-18}
\format{A GRanges object with length 2441.}
\usage{
data(hg19.refseq.transcripts)
}
\value{
A GRanges object.
}
\description{
This dataset contains a subset of the transcripts from the Refseq annotation
for mouse genome build hg19 See the ChIA-PET analysis vignette
(\code{vignettes(GenomicInteractions)}) for more information on how this
dataset was created.
}
\keyword{datasets}

