% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{is.pp}
\alias{InteractionHelpers}
\alias{is.cis}
\alias{is.cis,GenomicInteractions-method}
\alias{is.dd}
\alias{is.dd,GenomicInteractions-method}
\alias{is.dt}
\alias{is.dt,GenomicInteractions-method}
\alias{is.pd}
\alias{is.pd,GenomicInteractions-method}
\alias{is.pp}
\alias{is.pp,GenomicInteractions-method}
\alias{is.pt}
\alias{is.pt,GenomicInteractions-method}
\alias{is.trans}
\alias{is.trans,GenomicInteractions-method}
\alias{is.tt}
\alias{is.tt,GenomicInteractions-method}
\alias{isInteractionType}
\alias{isInteractionType,GenomicInteractions-method}
\title{Interaction Type Helpers}
\usage{
is.pp(GIObject)

is.pd(GIObject)

is.pt(GIObject)

is.dd(GIObject)

is.dt(GIObject)

is.tt(GIObject)

isInteractionType(GIObject, x, y)

is.trans(GIObject)

is.cis(GIObject)

\S4method{is.pp}{GenomicInteractions}(GIObject)

\S4method{is.pd}{GenomicInteractions}(GIObject)

\S4method{is.pt}{GenomicInteractions}(GIObject)

\S4method{is.dd}{GenomicInteractions}(GIObject)

\S4method{is.dt}{GenomicInteractions}(GIObject)

\S4method{is.tt}{GenomicInteractions}(GIObject)

\S4method{isInteractionType}{GenomicInteractions}(GIObject, x, y)

\S4method{is.trans}{GenomicInteractions}(GIObject)

\S4method{is.cis}{GenomicInteractions}(GIObject)
}
\arguments{
\item{GIObject}{A GenomicInteractions object}

\item{x,y}{Names of annotated node classes}
}
\value{
A logical vector
}
\description{
Functions to classify interactions within GenomicInteractions objects.
\itemize{
    \item "isInteractionType" takes two character arguments which are
          annotated node classes and returns interactions between them.
    \item "is.pp", "is.pd" etc. are bindings for common annotations:
    \describe{ \item{p}{promoter}
               \item{d}{distal}
               \item{t}{terminator} }
    \item "is.trans" & "is.cis" select trans-chromosomal and
          intra-chromosomal interactions, respectively }
}

