% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentData.R
\docType{data}
\name{mm9_refseq_promoters}
\alias{mm9_refseq_promoters}
\title{Mouse Refseq promoters from chr 14-15}
\format{A GRanges object with length 2441.}
\usage{
data(mm9_refseq_promoters)
}
\value{
A GRanges object.
}
\description{
This dataset contains a subset of the promoters from the Refseq annotation
for mouse genome build mm9. See the HiC analysis vignette 
(\code{vignettes(GenomicInteractions)}) for more information on how this 
dataset was created.
}
\keyword{datasets}

