% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\docType{methods}
\name{print,GenomicInteractions-method}
\alias{print,GenomicInteractions-method}
\title{Print function for GenomicInteractions}
\usage{
\S4method{print}{GenomicInteractions}(x)
}
\arguments{
\item{x}{GenomicInteractionsObject}
}
\value{
invisible(1)
}
\description{
Print function for GenomicInteractions
}

