% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{removeDups}
\alias{removeDups}
\title{Remove all but one occurences of a duplicated interaction}
\usage{
removeDups(GIObject)
}
\arguments{
\item{GIObject}{A GInteractions object.}
}
\value{
A GInteractions object that is a subset of the input object.
}
\description{
Removes all but the first occurence of a duplicated interaction (defined as
having identical coordinates for both anchors). N.B. this does not summarise
the total counts of all the duplicates. It is designed for removing potential
PCR duplicates after reading in .bam files.
}

