% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{resetAnnotations}
\alias{resetAnnotations}
\alias{resetAnnotations,GenomicInteractions-method}
\title{Reset annotations made to a GenomicInteractions object}
\usage{
resetAnnotations(GIObject)

\S4method{resetAnnotations}{GenomicInteractions}(GIObject)
}
\arguments{
\item{GIObject}{An annotated GenomicInteractions object}
}
\value{
invisible(1)
}
\description{
This function removes all annotations from a GenomicInteractions object by
deleting  all of the metadata columns associated with both anchors.
}

