% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,GenomicInteractions-method}
\alias{show,GenomicInteractions-method}
\title{Representation function for GenomicInteractions}
\usage{
\S4method{show}{GenomicInteractions}(object)
}
\arguments{
\item{object}{A GenomicInteractionsObject}
}
\value{
invisible(1)
}
\description{
Representation function for GenomicInteractions
}

