% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sum,GenomicInteractions-method}
\alias{sum,GenomicInteractions-method}
\title{Return the total number of interactions in a GenomicInteractions GIObject}
\usage{
\S4method{sum}{GenomicInteractions}(x)
}
\arguments{
\item{x}{GenomicInteractions GIObject}
}
\value{
The sum of the counts in GIObject
}
\description{
Return the total number of interactions in a GenomicInteractions GIObject
}

