% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viewPoints.R
\name{viewPoint}
\alias{viewPoint}
\title{Virtual 4C viewpoint}
\usage{
viewPoint(x, bait, region = NULL, ...)
}
\arguments{
\item{x}{A GenomicInteractions object.}

\item{bait}{A GRanges object describing bait regions.}

\item{region}{If present, a GenomicInteractions object specifying the
region to look for bait interactions in.}

\item{...}{additional arguments to findoverlaps}
}
\value{
A GenomicInteractions object.
}
\description{
This function creates a GenomicInteractions object representing interactions
originating at a given viewpoint ("bait"), or set of viewpoints. This is
similar to the idea of a virtual 4C experiment where you are interested in
interactions with a specific region.
}
\details{
The object returned has the "bait" as anchor one, and the interacting regions
as anchor two. By default this is genome wide. If you only want to consider
interactions within a certain distance around the bait, you can specify a
region to consider.

Multiple baits can be given, e.g. to find all interactions around promoters.

You may want to visualise the resulting interactions in a genome browser -
you can do this by creating coverage over anchor two of the object and exporting
as a wig or bedgraph file.
}
\examples{
\dontrun{
data(hic_data)
library(GenomicRanges)
pos <- GRanges(seqnames="chr5", ranges=IRanges(start=115938063, end=115941352))
region <- GRanges(seqnames="chr5", ranges=IRanges(start=115838063, end=116041352))
viewPoint(hic_data, pos, region)
}
}

