% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/viewPoints.R
\name{viewPoint}
\alias{viewPoint}
\title{Virtual 4C viewpoint}
\usage{
viewPoint(x, bait, region = NULL, ...)
}
\arguments{
\item{x}{A GenomicInteractions object.}

\item{bait}{A GRanges object describing bait regions.}

\item{region}{If present, a GenomicInteractions object specifying the
region to look for bait interactions in.}

\item{...}{additional arguments to findoverlaps}
}
\value{
A GenomicInteractions object.
}
\description{
Some explanation needed
}
\examples{
\dontrun{
data(hic_data)
library(GenomicRanges)
pos <- GRanges(seqnames="chr5", ranges=IRanges(start=115938063, end=115941352))
region <- GRanges(seqnames="chr5", ranges=IRanges(start=115838063, end=116041352))
viewPoint(hic_data, pos, region)
}
}

